requirements.txt
absl-py==2.2.1
attrs==25.3.0
cffi==1.17.1
contourpy==1.3.1
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.56.0
gaze-tracking==0.0.1
jax==0.5.3
jaxlib==0.5.3
kiwisolver==1.4.8
matplotlib==3.10.1
mediapipe==0.10.14
ml_dtypes==0.5.1
msvc_runtime==14.42.34433
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==4.25.6
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
scipy==1.15.2
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.1
tzdata==2025.2
